src/MCP/sales/requirements.txt
httpx
uvicorn
mcp==1.13.0
python-dotenv==1.0.0
numpy
pandas
pyarrow
databricks-sql-connector==4.1.2
azure-core==1.30.2
azure-identity==1.17.1
